kill only if present

+++ lib/utils.sh
@@ -163,10 +163,12 @@ function util_kill(){
         shift
     done
     local esckey="[${key:0:1}]${key:1}"
-    if [ -n "$ignore" ]; then
-        ps aux | grep $esckey | $ignore | sed -n 's/ \+/ /gp' | cut -d' ' -f2 | xargs kill -9 > /dev/null 2>&1
-    else
-        ps aux | grep $esckey | sed -n 's/ \+/ /gp' | cut -d' ' -f2 | xargs kill -9 > /dev/null 2>&1
+    if [ -n "$(ps aux | grep $esckey)" ]; then
+        if [ -n "$ignore" ]; then
+            ps aux | grep $esckey | $ignore | sed -n 's/ \+/ /gp' | cut -d' ' -f2 | xargs kill -9 > /dev/null 2>&1
+        else
+            ps aux | grep $esckey | sed -n 's/ \+/ /gp' | cut -d' ' -f2 | xargs kill -9 > /dev/null 2>&1
+        fi
     fi
 }
